+++ frontend/src/business/overmind/Api/state.ts
@@ -1,5 +1,15 @@
-import { ApiState } from "../../models/State";
+export interface ApiState {
+  response: {
+    statusText?: string;
+    message: string;
+    status: "success" | "error" | undefined;
+  };
+}
 
 export const state: ApiState = {
-  response: { statusText: "", message: "", status: undefined },
+  response: {
+    statusText: "",
+    message: "",
+    status: undefined,
+  },
 };
