+++ frontend/src/business/overmind/App/state.ts
@@ -1,4 +1,6 @@
-import { AppState } from "../../models/State";
+export interface AppState {
+  isLoading: boolean;
+}
 
 export const state: AppState = {
   isLoading: false,
