+++ frontend/src/business/overmind/Author/state.ts
@@ -1,5 +1,20 @@
 import { derived } from "overmind";
-import { AuthorState } from "../../models/State";
+import { ApiResponse } from "../../models/Api";
+import { Author, AuthorStatsPage } from "../../models/Author";
+import { BaseState } from "../../models/State";
+
+export interface AuthorState extends BaseState<Author> {
+  statsPage: AuthorStatsPage | null;
+  statsPageApi: ApiResponse<AuthorStatsPage> | null;
+  ui: {
+    table: {
+      limit: number;
+      page: number;
+      queryString: string;
+      showAll: boolean;
+    };
+  };
+}
 
 export const state: AuthorState = {
   getAllApi: null,
