+++ frontend/src/business/overmind/Bookworm/state.ts
@@ -1,5 +1,19 @@
 import { derived } from "overmind";
-import { BookwormState } from "../../models/State";
+import { Bookworm } from "../../models/Bookworm";
+import { BaseState } from "../../models/State";
+import { UserApi } from "../../models/User";
+
+export interface BookwormState extends BaseState<Bookworm> {
+  bookwormDetailsApi: Omit<UserApi, "token"> | null;
+  ui: {
+    table: {
+      limit: number;
+      page: number;
+      queryString: string;
+      showAll: boolean;
+    };
+  };
+}
 
 export const state: BookwormState = {
   getAllApi: null,
