+++ frontend/src/business/overmind/User/state.ts
@@ -1,4 +1,8 @@
-import { UserState } from "../../models/State";
+import { User } from "../../models/User";
+
+export interface UserState {
+  user: User | null;
+}
 
 export const state: UserState = {
   user: null,
